+++ streams-bootstrap-core/src/test/java/com/bakdata/kafka/admin/TopicsClientClusterTest.java
@@ -29,6 +29,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 
 import com.bakdata.kafka.ApacheKafkaContainerCluster;
+import com.bakdata.kafka.admin.TopicsClient.TopicClient;
 import java.time.Duration;
 import java.util.Map;
 import org.apache.kafka.clients.admin.AdminClientConfig;
@@ -38,29 +39,32 @@ import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
 @Testcontainers
-class TopicClientClusterTest {
+class TopicsClientClusterTest {
 
     private static final Duration CLIENT_TIMEOUT = Duration.ofSeconds(10L);
     @Container
-    private final ApacheKafkaContainerCluster kafkaCluster = new ApacheKafkaContainerCluster(AppInfoParser.getVersion(), 3, 2);
+    private final ApacheKafkaContainerCluster kafkaCluster =
+            new ApacheKafkaContainerCluster(AppInfoParser.getVersion(), 3, 2);
 
     @Test
     void shouldCreateTopicWithReplication() {
-        try (final TopicClient client = this.createClient()) {
-            assertThat(client.exists("topic")).isFalse();
+        try (final AdminClientX admin = this.createAdminClient()) {
+            final TopicsClient client = admin.topics();
+            final TopicClient topic = client.topic("topic");
+            assertThat(topic.exists()).isFalse();
             final TopicSettings settings = TopicSettings.builder()
                     .partitions(5)
                     .replicationFactor((short) 2)
                     .build();
-            client.createTopic("topic", settings, emptyMap());
+            topic.create(settings, emptyMap());
             // topic needs to be propagated to all KRaft controllers
             // topic creation only verifies existence on one controller
             await()
                     .pollDelay(Duration.ofSeconds(1L))
                     .atMost(Duration.ofSeconds(20L))
                     .untilAsserted(() -> {
-                        assertThat(client.exists("topic")).isTrue();
-                        assertThat(client.describe("topic"))
+                        assertThat(topic.exists()).isTrue();
+                        assertThat(topic.getSettings())
                                 .hasValueSatisfying(info -> {
                                     assertThat(info.getReplicationFactor()).isEqualTo((short) 2);
                                     assertThat(info.getPartitions()).isEqualTo(5);
@@ -69,10 +73,10 @@ class TopicClientClusterTest {
         }
     }
 
-    private TopicClient createClient() {
+    private AdminClientX createAdminClient() {
         final String brokerList = this.kafkaCluster.getBootstrapServers();
         final Map<String, Object> config = Map.of(AdminClientConfig.BOOTSTRAP_SERVERS_CONFIG, brokerList);
-        return TopicClient.create(config, CLIENT_TIMEOUT);
+        return AdminClientX.create(config, CLIENT_TIMEOUT);
     }
 
 }