+++ streams-bootstrap-core/src/test/java/com/bakdata/kafka/producer/ProducerCleanUpRunnerTest.java
@@ -98,7 +98,7 @@ class ProducerCleanUpRunnerTest extends KafkaTest {
     void shouldDeleteTopic() {
         try (final ConfiguredProducerApp<ProducerApp> app = createStringApplication();
                 final ExecutableProducerApp<ProducerApp> executableApp = app
-                        .withRuntimeConfiguration(this.createConfigWithoutSchemaRegistry())) {
+                        .withRuntimeConfiguration(this.createConfig())) {
             run(executableApp);
 
             final List<KeyValue<String, String>> output = this.readOutputTopic(app.getTopics().getOutputTopic());
@@ -108,7 +108,7 @@ class ProducerCleanUpRunnerTest extends KafkaTest {
             clean(executableApp);
 
             try (final AdminClientX admin = this.newTestClient().admin()) {
-                this.softly.assertThat(admin.getTopicClient().exists(app.getTopics().getOutputTopic()))
+                this.softly.assertThat(admin.topics().topic(app.getTopics().getOutputTopic()).exists())
                         .as("Output topic is deleted")
                         .isFalse();
             }
@@ -119,7 +119,7 @@ class ProducerCleanUpRunnerTest extends KafkaTest {
     void shouldDeleteValueSchema() throws IOException, RestClientException {
         try (final ConfiguredProducerApp<ProducerApp> app = createAvroValueApplication();
                 final ExecutableProducerApp<ProducerApp> executableApp = app
-                        .withRuntimeConfiguration(this.createConfig());
+                        .withRuntimeConfiguration(this.createConfigWithSchemaRegistry());
                 final SchemaRegistryClient client = this.getSchemaRegistryClient()) {
             run(executableApp);
 
@@ -136,7 +136,7 @@ class ProducerCleanUpRunnerTest extends KafkaTest {
     void shouldDeleteKeySchema() throws IOException, RestClientException {
         try (final ConfiguredProducerApp<ProducerApp> app = createAvroKeyApplication();
                 final ExecutableProducerApp<ProducerApp> executableApp = app
-                        .withRuntimeConfiguration(this.createConfig());
+                        .withRuntimeConfiguration(this.createConfigWithSchemaRegistry());
                 final SchemaRegistryClient client = this.getSchemaRegistryClient()) {
             run(executableApp);